requirements.txt
Django==4.1.3
psycopg2-binary>=2.8
djangorestframework>=3.9.4
PyJWT==2.6.0